Add missing no cleanup flags
Make the rest of the tests have the same no cleanup flag that the
production test has. This will be used in the future to ensure
upgrading those features doesn't break.

+++ examples/tornjak/run-tests.sh
@@ -12,9 +12,22 @@ source "${TESTDIR}/common.sh"
 helm_install=(helm upgrade --install --create-namespace)
 ns=spire-system
 
+CLEANUP=1
+
+for i in "$@"; do
+  case $i in
+    -c)
+      CLEANUP=0
+      shift # past argument=value
+      ;;
+  esac
+done
+
 teardown() {
-  helm uninstall --namespace "${ns}" spire 2>/dev/null || true
-  kubectl delete ns "${ns}" 2>/dev/null || true
+  if [ "${CLEANUP}" -eq 1 ]; then
+    helm uninstall --namespace "${ns}" spire 2>/dev/null || true
+    kubectl delete ns "${ns}" 2>/dev/null || true
+  fi
 }
 
 trap 'trap - SIGTERM && teardown' SIGINT SIGTERM EXIT
